move validation logic to rule.rs
